pyproject.toml
[build-system]
requires = [
    "setuptools>=80.0.0",
    "wheel",
    "packaging>=24.0"
]
build-backend = "setuptools.build_meta"

[project]
name = "spex-tools"
version = "0.3.1059"
description = "Comprehensive spatial transcriptomics analysis library with advanced segmentation, clustering, and spatial analysis tools"
readme = "README.md"
authors = [{name = "Aleksandr Prilipko"}]
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "aicsimageio==4.10.0",
    "imagecodecs>=2020.5.30",
    "fsspec>=2022.8.0",
    "nd2>=0.2.0",
    "mrc>=0.2.0",
    "bfio==2.3.5",
    "imageio-ffmpeg>=0.4",
    "Pillow!=8.3.0,<9,>=8.2.0",
    "ome-types==0.6.1",
    "tifffile>=2022.8.12,<2023.3.15",
    "numpy>=1.25.2,<1.26",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-extra-types==2.10.5",
    "typing_extensions>=4.7.1",
    "bioformats-jar",
    "jpype1",
    "scikit-learn==1.3.2",
    "scikit-image>=0.19.3,<0.20.0",
    "scikit-misc>=0.1.5",
    "dask==2021.11.1",
    "keras==2.14.0",
    "tensorflow==2.14.0",
    "stardist==0.8.5",
    "opencv-python==4.5.5.64",
    "opencv-python-headless==4.5.5.64",
    "csbdeep==0.7.4",
    "protobuf>=3.9.2,<3.20.4",
    "numexpr==2.8.4",
    "packaging",
    "cellpose==0.6.5",
    "anndata==0.8.0",
    "phenograph==1.5.7",
    "scanpy==1.9.4",
    "leidenalg==0.9.1",
    "louvain==0.8.2",
    "squidpy==1.2.2",
    "scvi-tools[scanpy]==1.1.6.post2",
    "jax==0.4.12",
    "jaxlib==0.4.12",
    "scipy>=1.11.1,<1.13",
    "pegasuspy==1.10.2",
    "treescope>=0.1.9",
    "fastapi==0.115.14",
    "pegasusio",
    "pandas==2.2.0",
    "decoupler==2.1.0",
    "jaxtyping==0.3.2",
    "flax==0.7.4",
    "zarr==2.18.2",
    "cython==3.1.2",
    "requests==2.32.4",
    "cmake==4.0.3",
    "pyarrow==20.0.0",
    "numcodecs==0.12.1"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
    "pre-commit"
]
docs = [
    "mkdocs-material",
    "mkdocstrings[python]",
    "mkdocs-jupyter"
]

[tool.setuptools]
packages = { find = {} }

[project.urls]
Homepage = "https://genentech.github.io/spex-tools/"
Repository = "https://github.com/genentech/spex-tools"
Documentation = "https://genentech.github.io/spex-tools/"

[tool.setuptools.package-data]
"spex.resources" = ["*.parquet"]